requirements.txt
attrs>=19.1
cryptoparser>=0.7.1
cryptolyzer>=0.7.1
urllib3
